import issue fixed for /node_js components
