+++ src/Router/AppRouter.js
@@ -20,7 +20,7 @@ import Fundamentals from "../pages/fundamentals/Fundamentals";
 import Bitwise from "../pages/javascript_core/bitwise";
 import Ternary from "../pages/frontend_frameworks/ternary";
 import Sixth from "../pages/techtrends";
-import Switchjs from "../pages/switch";
+import Switchjs from "../pages/node_js/switch";
 import Fivth from "../pages/loop";
 import Fundamentals2 from "../pages/fundamentals/12";
 import Fundamentals3 from "../pages/fundamentals/13";
@@ -49,15 +49,15 @@ import Ternary7 from "../pages/frontend_frameworks/37";
 import Ternary8 from "../pages/frontend_frameworks/38";
 import Ternary9 from "../pages/frontend_frameworks/39";
 import Ternary10 from "../pages/frontend_frameworks/310";
-import Switchjs2 from "../pages/42";
-import Switchjs3 from "../pages/43";
-import Switchjs4 from "../pages/44";
-import Switchjs5 from "../pages/45";
-import Switchjs6 from "../pages/46";
-import Switchjs7 from "../pages/47";
-import Switchjs8 from "../pages/48";
-import Switchjs9 from "../pages/49";
-import Switchjs10 from "../pages/410";
+import Switchjs2 from "../pages/node_js/42";
+import Switchjs3 from "../pages/node_js/43";
+import Switchjs4 from "../pages/node_js/44";
+import Switchjs5 from "../pages/node_js/45";
+import Switchjs6 from "../pages/node_js/46";
+import Switchjs7 from "../pages/node_js/47";
+import Switchjs8 from "../pages/node_js/48";
+import Switchjs9 from "../pages/node_js/49";
+import Switchjs10 from "../pages/node_js/410";
 import Fivth2 from "../pages/52";
 import Fivth3 from "../pages/53";
 import Fivth4 from "../pages/54";
